Experimental: Moves structures into separate files
